+++ includes/Hooks.php
@@ -3,6 +3,7 @@
 namespace MediaWiki\Extension\DynamicPageList3;
 
 use DatabaseUpdater;
+use ExtensionRegistry;
 use MediaWiki\Extension\DynamicPageList3\Maintenance\CreateTemplate;
 use MediaWiki\Extension\DynamicPageList3\Maintenance\CreateView;
 use Parser;
@@ -82,7 +83,7 @@ class Hooks {
 		static $version = null;
 
 		if ( $version === null ) {
-			$version = json_decode( file_get_contents( __DIR__ . '/../extension.json' ) )->version;
+			$version = ExtensionRegistry::getInstance()->getAllThings()['DynamicPageList3']['version'];
 		}
 
 		return $version;
@@ -370,13 +371,19 @@ class Hooks {
 		// convert \n to a real newline character
 		$repl = str_replace( '\n', "\n", $repl );
 
-		# replace
+		// replace
 		if ( !self::isRegexp( $pat ) ) {
 			$pat = '`' . str_replace( '`', '\`', $pat ) . '`';
 		}
 
+		// Check for buffer overflow
+		if ( strlen( $pat ) > 1000 ) {
+			return '';
+		}
+
 		// phpcs:ignore Generic.PHP.NoSilencedErrors.Discouraged
 		return @preg_replace( $pat, $repl, $text );
+		// @phan-suppress-previous-line SecurityCheck-ReDoS
 	}
 
 	/**