+++ CHANGELOG.md
@@ -810,3 +810,6 @@ Many thanks to GreenReaper on GitHub for reporting and finding issues with core 
 
 # Version 3.5.2
 * Added support for PHP 8.0 and PHP 8.1
+* Enabled the `SecurityCheck-ReDoS` phan check
+* Added buffer overflow check for `dplreplace`
+* Fixed `usedby` queries returning all templates